requirements.txt
openai
langchain
google-search-results
streamlit
python-dotenv
langchain_groq 